+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - Updated to Rust v1.85 and 2024 edition.
 
+#### 🐞 Fixes
+
+- Fixed an issue where importing `Result` would break our generated macro code.
+
 #### ⚙️ Internal
 
 - Updated dependencies.